Refactor Views to pages directory
