+++ frontend/package.json
@@ -14,6 +14,7 @@
     },
     "dependencies": {
         "@apollo/client": "^3.3.7",
+        "@chakra-ui/icons": "^1.0.6",
         "@chakra-ui/react": "^1.3.4",
         "@emotion/react": "^11.1.4",
         "@emotion/styled": "^11.0.0",
