+++ frontend/pages/search.tsx
@@ -1,7 +1,19 @@
 import React from "react";
-import SearchView from "@components/pages/SearchView";
 import { NextPage } from "next";
+import NavBar from "@components/molecules/NavBar";
+import Song from "./song";
+import { Flex } from "@chakra-ui/react";
+import SearchInput from "@components/atoms/SearchInput";
+import Header from "@components/molecules/Header";
 
-const Search: NextPage = () => <SearchView />;
+const Search: NextPage = () => (
+    <Flex flexDir="column" w="100%">
+        <Header color="#FFD687">
+            <SearchInput />
+        </Header>
+        <Song />
+        <NavBar />
+    </Flex>
+);
 
 export default Search;
