--- frontend/src/components/pages/SearchView.tsx
@@ -1,12 +1,0 @@
-import React from "react";
-import NavBar from "../molecules/NavBar";
-import SongView from "./SongView";
-
-const SearchView = () => (
-    <div>
-        <SongView />
-        <NavBar />
-    </div>
-);
-
-export default SearchView;
